update crud api, auth login, add more attribute owner_id in organizations

+++ src/main/java/com/example/taskmanagement_backend/dtos/UserProfileDto/CreateUserProfileRequestDto.java
@@ -19,7 +19,6 @@ public class CreateUserProfileRequestDto {
 
     private String status;
 
-
     @NotNull
-    private Integer userId;
+    private Long userId;
 }
